server/requirements/base.txt
coreapi==2.3.3
Django==1.11.22
django-anymail==3.0
django-cors-headers==2.1.0
django-debug-toolbar==1.8
django-enumfields==0.10.1
django-extensions==1.9.1
django-filter==1.1.0
django-model-utils==3.1.1
django-phonenumber-field==2.0.0
django-suit~=0.2.28
django-suit-daterange-filter==0.0.6
django-templated-email==2.2.0
djangorestframework==3.10.2
dj-database-url==0.4.2
Faker~=2.0.0
faker-e164~=0.1.0
flake8==3.3.0
ipdb==0.10.3
isort==4.2.5
psycopg2~=2.8.3
pytest==3.2.3
pytest-django==3.1.2
pytest-env==0.6.2
Pillow==6.2.0
rules~=2.1
twilio==6.14.10
uWSGI==2.0.15
whitenoise==3.3.1
graphene-django~=2.1.8
